backend/requirements.txt
Django==5.1
djangorestframework==3.15.2
drf-spectacular==0.28.0
psycopg2-binary==2.9.10
django-cors-headers==4.6.0
gunicorn==23.0.0
Pillow==11.1.0
